+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [unreleased]
 
 - Add a new configuration flag that allows displaying without trailing whitespace. The whitespace token is also configurable.
+- Add `show_print_commands` method that converts a list of tokens into its wadler printing commands.
 
 ## v0.9.6 (25-11-2024)